move filter by collections to own component
